Update main to be more readable.
Functionality is the same, with the major change being comments, and
also breaking up stream processing into its own function so that instead
of setting a response string var to literal "Ok: ..." and "Err: ..."
string, it returns Result enum values that get cast as the result
strings in a single place later on. Because rust, and woo!